[2.18.1-SNAPSHOT][fix] - fix month age filter - Ref gestion-de-projet#1447
